+++ src/routes/user.py
@@ -4,7 +4,7 @@ from src.auth import get_current_user, hash_password
 from src.crud.user import (check_username_exists, create_user, delete_user,
                            get_user_by_id, get_user_by_username, update_user)
 from src.database import get_database
-from src.models.user import UserCreate, UserResponse, UserUpdate
+from src.models.user import UserCreate, UserResponse, UserUpdate, Role
 
 router = APIRouter(prefix="/user", tags=["user"])
 
@@ -13,13 +13,15 @@ router = APIRouter(prefix="/user", tags=["user"])
 async def user_create_route(
     user: UserCreate, user_data=Depends(get_current_user), db=Depends(get_database)
 ):
-    existing_username = check_username_exists(user.username)
-    if existing_username:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="User with this username already exists",
-        )
-    return await create_user(user, hash_password(user.password), db)
+    if user_data["role"] == Role.ADMIN:
+        existing_username = await check_username_exists(user.username, db)
+        if existing_username:
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail="User with this username already exists",
+            )
+        return await create_user(user, hash_password(user.password), db)
+    raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You cannot create a user as a member")
 
 
 @router.get("/id/{user_id}", response_model=UserResponse)
@@ -45,21 +47,25 @@ async def user_update_route(
     user_data=Depends(get_current_user),
     db=Depends(get_database),
 ):
-    if user.username:
-        existing_username = check_username_exists(user.username)
-        if existing_username:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="User with this username already exists",
-            )
+    if user_data["role"] == Role.ADMIN or id == user_data["id"]:
+        if user.username:
+            existing_username = await check_username_exists(user.username, db)
+            if existing_username:
+                raise HTTPException(
+                    status_code=status.HTTP_400_BAD_REQUEST,
+                    detail="User with this username already exists",
+                )
 
-    return await update_user(id, user, db)
+        return await update_user(id, user, db)
+    raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You can't edit other users info")
 
 
 @router.delete("/{username}", status_code=status.HTTP_204_NO_CONTENT)
 async def user_delete_route(
     username: str, user_data=Depends(get_current_user), db=Depends(get_database)
 ):
-    deleted_user = await delete_user(username, db)
-    if not deleted_user:
-        raise HTTPException(status.HTTP_404_NOT_FOUND, detail="User not found")
+    if user_data["role"] == Role.ADMIN or id == user_data["id"]:
+        deleted_user = await delete_user(username, db)
+        if not deleted_user:
+            raise HTTPException(status.HTTP_404_NOT_FOUND, detail="User not found")
+    raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You can't delete other users")
